# helix-core/src/openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: Helix ID
  version: "0.1.0"
  description: Helix ID agent identity and trust infrastructure API.
paths:
...
  /v1/dids:
    post:
      summary: Create a DID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDIDRequest"
      responses:
        "201":
          description: DID created
  /v1/dids/{did}:
    get:
      summary: Resolve a DID
      parameters:
        - in: path
          name: did
          required: true
          schema:
            type: string
        - in: query
          name: live
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: DID resolved
  /v1/dids/{did}/services:
    post:
      summary: Add a DID service endpoint
      parameters:
        - in: path
          name: did
          required: true
          schema:
            type: string
      responses:
        "200":
          description: DID updated
  /v1/dids/{did}/services/{endpointId}:
    delete:
      summary: Remove a DID service endpoint
      parameters:
        - in: path
          name: did
          required: true
          schema:
            type: string
        - in: path
          name: endpointId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: DID updated
  /v1/dids/{did}/deactivate:
    post:
      summary: Deactivate a DID
      parameters:
        - in: path
          name: did
          required: true
          schema:
            type: string
      responses:
        "204":
          description: DID deactivated
...
components:
  schemas:
    CreateDIDRequest:
      type: object
      required: [publicKeyHex, subjectType]
      properties:
        publicKeyHex:
          type: string
        subjectType:
          type: string
          enum: [agent, user]
        domains:
          type: array
          items:
            type: string
...